+++ docs/examples/bucket_group/bucket_group_trio.py
@@ -1,13 +1,11 @@
 from trio import run, sleep
 from trio.lowlevel import checkpoint
-from rate_control import BucketGroup, Duration, FixedWindowCounter, RateLimit, RateLimiter
+from rate_control import Duration, FixedWindowCounter, RateLimit, RateLimiter
 
 async def main() -> None:
     first_bucket = FixedWindowCounter(2, Duration.SECOND)
     second_bucket = FixedWindowCounter(3, 2 * Duration.SECOND)
-    async with BucketGroup(first_bucket, second_bucket) as bucket_group:
-        rate_limiter = RateLimiter(bucket_group)
-
+    async with RateLimiter(first_bucket, second_bucket) as rate_limiter:
         async with rate_limiter.request():
             print('First request passes')
         async with rate_limiter.request():