+++ docs/examples/scheduler/scheduler_trio.py
@@ -6,9 +6,8 @@ async def request_print(controller: RateController, start_time: float) -> None:
         print(f'Elapsed: {current_time() - start_time :.1f} seconds')
 
 async def main() -> None:
-    async with FixedWindowCounter(capacity=2, duration=Duration.SECOND) as bucket, \
-            Scheduler(bucket) as scheduler, \
-            open_nursery() as nursery:
+    bucket = FixedWindowCounter(capacity=2, duration=Duration.SECOND)
+    async with Scheduler(bucket) as scheduler, open_nursery() as nursery:
         for _ in range(3):
             nursery.start_soon(request_print, scheduler, current_time())